+++ lava_android_test/test_definitions/busybox.py
@@ -17,6 +17,15 @@
 #
 # You should have received a copy of the GNU General Public License
 # along with this program.  If not, see <http://www.gnu.org/licenses/>.
+"""
+This is a simple test to ensure busybox is installed on an Android build and
+can execute some basic busybox commands
+
+**URL:**  http://www.busybox.net/
+
+**Default Options:** None
+"""
+
 import os
 import lava_android_test.testdef
 from lava_android_test.config import get_config