+++ lava_android_test/test_definitions/mmtest.py
@@ -17,8 +17,17 @@
 #
 # You should have received a copy of the GNU General Public License
 # along with this program.  If not, see <http://www.gnu.org/licenses/>.
+
+"""
+Performs a test of multimedia functionality in Android by playing a variety
+of different multimedia formats on Android.
+
+**URL:** http://samplemedia.linaro.org/
+
+**Default options:** None
+"""
+
 import os
-import re
 import lava_android_test.testdef
 from lava_android_test.utils import get_local_name
 from lava_android_test.config import get_config
@@ -42,59 +51,12 @@ RUN_ADB_SHELL_STEPS = ['am instrument -r -e targetDir %s \
      % test_files_target_path,
     'rm -r %s' % (test_files_target_path)]
 
-
-class MMTestTestParser(lava_android_test.testdef.AndroidTestParser):
-
-    def parse(self, result_filename='stdout.log', output_filename='stdout.log',
-               test_name=test_name):
-        """Parse test output to gather results
-        Use the pattern specified when the class was instantiated to look
-        through the results line-by-line and find lines that match it.
-        Results are then stored in self.results.  If a fixupdict was supplied
-        it is used to convert test result strings to a standard format.
-        """
-        pat_test = re.compile(
-            r'^\s*INSTRUMENTATION_STATUS:\s*test=(?P<test_case_id>.+)\s*$')
-        pat_status_code = re.compile(
-            r'^\s*INSTRUMENTATION_STATUS_CODE:\s*(?P<status_code>[\d-]+)\s*$')
-        data = {}
-        with open(output_filename, 'r') as stream:
-            for lineno, line in enumerate(stream, 1):
-                match = pat_test.search(line)
-                if match:
-                    data['test_case_id'] = match.group('test_case_id')
-                    continue
-
-                match = pat_status_code.search(line)
-                if match:
-                    status_code = match.group('status_code')
-                    if status_code == '1':
-                        # test case started
-                        data = {}
-                    elif data['test_case_id']:
-                        if status_code == '0':
-                            data['result'] = 'pass'
-                        else:
-                            data['result'] = 'fail'
-                        data["log_filename"] = result_filename
-                        data["log_lineno"] = lineno
-                        self.results['test_results'].append(data)
-                        data = {}
-                    continue
-
-        if self.fixupdict:
-            self.fixresults(self.fixupdict)
-        if self.appendall:
-            self.appendtoall(self.appendall)
-        self.fixmeasurements()
-        self.fixids()
-
 inst = lava_android_test.testdef.AndroidTestInstaller()
 run = lava_android_test.testdef.AndroidTestRunner(
                                 steps_host_pre=RUN_STEPS_HOST_PRE,
                                 steps_adb_pre=RUN_STEPS_ADB_PRE,
                                 adbshell_steps=RUN_ADB_SHELL_STEPS)
-parser = MMTestTestParser()
+parser = lava_android_test.testdef.AndroidInstrumentTestParser()
 testobj = lava_android_test.testdef.AndroidTest(testname=test_name,
                                                 installer=inst,
                                                 runner=run,