+++ lava_android_test/test_definitions/pm-qa.py
@@ -17,14 +17,24 @@
 #
 # You should have received a copy of the GNU General Public License
 # along with this program.  If not, see <http://www.gnu.org/licenses/>.
-
+import os
 import lava_android_test.testdef
+from lava_android_test.config import get_config
 
-test_name = 'memtester'
+test_name = 'pm-qa'
+config = get_config()
+curdir = os.path.realpath(os.path.dirname(__file__))
+test_sh_name = 'pm-qa.sh'
+test_sh_path = os.path.join(curdir, test_name, test_sh_name)
+test_sh_android_path = os.path.join(config.installdir_android,
+                                    test_name, test_sh_name)
 
-INSTALL_STEPS_ADB_PRE = []
-ADB_SHELL_STEPS = ['memtester 1M 1']
-PATTERN = "^\s*(?P<test_case_id>.*?)\s*:\s*(?P<result>\w+)\s*$"
+INSTALL_STEPS_ADB_PRE = ['push %s %s ' % (test_sh_path,
+                                          test_sh_android_path),
+                          'shell chmod 777 %s' % test_sh_android_path]
+
+ADB_SHELL_STEPS = [test_sh_android_path]
+PATTERN = "^\s*(?P<test_case_id>\w+)=(?P<result>\w+)\s*$"
 
 inst = lava_android_test.testdef.AndroidTestInstaller(
                                 steps_adb_pre=INSTALL_STEPS_ADB_PRE)