+++ lava_android_test/test_definitions/skia.py
@@ -17,6 +17,15 @@
 #
 # You should have received a copy of the GNU General Public License
 # along with this program.  If not, see <http://www.gnu.org/licenses/>.
+
+"""
+Runs the skia benchmark to test 2D graphics performance.
+
+**URL:** https://sites.google.com/site/skiadocs/
+
+**Default options:** None
+"""
+
 import re
 import lava_android_test.testdef
 
@@ -26,10 +35,10 @@ INSTALL_STEPS_ADB_PRE = []
 # Skia can do many more benchmarks, but it becomes almost too much data
 # to make a nice chart for. The -match limits the ones we run
 ADB_SHELL_STEPS = ['logcat -c',
-                   'skia_bench -repeat 1000 -timers w -config 565 -match bitmap',
-                   'skia_bench -repeat 1000 -timers w -config 565 -match rects',
-                   'skia_bench -repeat 1000 -timers w -config 565 -match repeat',
-                   'logcat -d -s "skia:*"']
+                 'skia_bench -repeat 1000 -timers w -config 565 -match bitmap',
+                 'skia_bench -repeat 1000 -timers w -config 565 -match rects',
+                 'skia_bench -repeat 1000 -timers w -config 565 -match repeat',
+                  'logcat -d -s "skia:*"']
 
 
 class SkiaTestParser(lava_android_test.testdef.AndroidTestParser):